OBPIH-5920 Improvements and fixes after moving 0.8.23 into 0.9.0 - Moved new Config.groovy configurations to the runtime.groovy - Moved some files to proper directory - Moved RequisitionStatusTransitionEventService to proper package - Replaced transactional = true with [at]Transactional - Replaced old ConfigurationHolder and ApplicationHolder with grails.util.Holders
