+++ grails-app/services/org/pih/warehouse/requisition/RequisitionStatusTransitionEventService.groovy
@@ -7,7 +7,7 @@
 * the terms of this license.
 * You must not remove this notice, or any other, from this software.
 **/
-package org.pih.warehouse.inventory
+package org.pih.warehouse.requisition
 
 import org.pih.warehouse.report.NotificationService
 import org.pih.warehouse.requisition.RequisitionStatusTransitionEvent
